requirements.txt
asgiref==3.8.1
Django==5.1.5
sqlparse==0.5.3
Pillow>=10.0.0
psycopg>=3.0.0
gunicorn==20.1.0
django-cors-headers
djangorestframework==3.14.0
